# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.UserDomain:
    properties:
      createdAt:
        type: string
      domainId:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/models.UserDomainRole'
        description: '0: read only, 1: read write, 2: manager, 3: owner'
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  models.UserDomainRole:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-comments:
      Manager: add R/W user, delete R/W user, R/W
      ReadWrite: manage records, but not others
    x-enum-varnames:
    - ReadOnly
    - ReadWrite
    - Manager
    - Owner
...
  web.Domain:
    properties:
      data: {}
      errors:
        type: string
      status:
        type: integer
    type: object
...
  web.DomainUser:
    properties:
      role:
        $ref: '#/definitions/models.UserDomainRole'
      user_id:
        type: integer
    type: object
  web.DomainUserDetail:
    properties:
      domain_id:
        type: integer
      domain_name:
        type: string
      email:
        type: string
      role:
        $ref: '#/definitions/models.UserDomainRole'
      user_id:
        type: integer
      username:
        type: string
    type: object
  web.User:
    properties:
      data: {}
      error:
        type: string
      status:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: makiras.x@outlook.com
    name: domain0
  description: Domain0 API
  license:
    name: MPL(mozilla public license)-2.0
    url: https://www.mozilla.org/en-US/MPL/2.0/
  title: Domain0 API
  version: 0.0.1
paths:
...
  /api/v1/domain/{id}/user:
    get:
      consumes:
      - application/json
      description: |-
        List UserDomain Relation
        user must have manager permission to domain or be admin
      parameters:
      - description: domain id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.DomainUserDetail'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
      summary: List UserDomain Relation
      tags:
      - domain
    post:
      consumes:
      - application/json
      description: |-
        Create UserDomain Relation
        user must have manager permission to domain or be admin
        user cant create permission higher than himself
      parameters:
      - description: domain id
        in: path
        name: id
        required: true
        type: string
      - description: userRole
        in: body
        name: userRole
        required: true
        schema:
          $ref: '#/definitions/web.DomainUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  $ref: '#/definitions/models.UserDomain'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
      summary: Create UserDomain Relation
      tags:
      - domain
  /api/v1/domain/{id}/user/{uid}:
    delete:
      consumes:
      - application/json
      description: |-
        Delete UserDomain Relation **(no update, just delete and create)**
        user must have manager permission to domain or be admin
        user cant delete permission higher than himself
      parameters:
      - description: domain id
        in: path
        name: id
        required: true
        type: string
      - description: user id
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/web.Domain'
            - properties:
                data:
                  type: integer
              type: object
      summary: Delete UserDomain Relation
      tags:
      - domain
...
  /api/v1/user/login:
    post:
      description: |-
        login api
        user can login with email or stu_id(Not implemented)
      parameters:
      - description: user email or stu_id
        in: formData
        name: user
        required: true
        type: string
      - description: user password
        in: formData
        name: pass
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
      summary: login
      tags:
      - user
  /api/v1/user/register:
    post:
      description: |-
        register api
        user can register with email
      parameters:
      - description: user email
        in: formData
        name: email
        required: true
        type: string
      - description: user password
        in: formData
        name: pass
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/web.User'
            - properties:
                data:
                  type: integer
              type: object
      summary: register
schemes:
- http
- https
swagger: "2.0"